common/pyproject.toml
[tool.poetry]
name = "common"
version = "0.1.0"
description = "Common utility package"
authors = ["mbm60-max <mbm60@bath.ac.uk>"]
readme = "README.md"
packages = [{include = "common"}]

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^2.1.0"
setuptools = "^80.9.0"

[tool.poetry.group.dev.dependencies]
black = ">=25.11.0,<26.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
